# perf: enable SWC `disableAllLints` by default to reduce overhead