--- a/packages/rspack-test-tools/tests/diagnosticsCases/config/builtin-swc-loader/stats.err
+++ b/packages/rspack-test-tools/tests/diagnosticsCases/config/builtin-swc-loader/stats.err
@@ -6,5 +6,5 @@ ERROR in × failed to parse builtin:swc-loader options
    ·             ▲
    ·             ╰── unknown field `parser2`, expected one of `assumptions`, `parser`, `transform`, `externalHelpers`, `target`, `loose`, `keepClassNames`, `baseUrl`, `paths`, `minify`, `experimental`, `lints`, `preserveAllComments`, `output` at line 3 column 13
  4 │       "syntax": "ecmascript"
- 5 │     }
+ 5 │     },
    ╰────