--- a/packages/rspack/src/builtin-loader/swc/types.ts
+++ b/packages/rspack/src/builtin-loader/swc/types.ts
@@ -25,9 +25,5 @@ export type SwcLoaderOptions = Config & {
 	 */
 	rspackExperiments?: {
 		import?: PluginImportOptions;
-		/**
-		 * @deprecated use `import` instead
-		 */
-		pluginImport?: PluginImportOptions;
 	};
 };